% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_svg.R
\name{inline_svg_file}
\alias{inline_svg_file}
\title{Insert SVG file as inline content}
\usage{
inline_svg_file(
  path,
  id = NULL,
  role = NULL,
  alt_title = NULL,
  alt_description = NULL,
  source_note = NULL,
  strip_ns = TRUE,
  dimensions = c("preserve", "strip", "replace"),
  width = NULL,
  height = NULL,
  units = NULL
)
}
\arguments{
\item{path}{Path to the SVG file}

\item{id}{An id value for the SVG element within the HTML document (optional)}

\item{role}{The ARIA role for the SVG element (optional)}

\item{alt_title}{A short title for assistive technology users (optional)}

\item{alt_description}{A longer description for assistive technology users
(optional)}

\item{source_note}{Text to include in a containing \verb{<div>} as a source or
reference note below the image.}

\item{strip_ns}{Whether to remove the XML DTD}

\item{dimensions}{Whether to 'preserve' dimensions defined in the SVG tag,
to 'strip' them or 'replace' them with provided values}

\item{width}{The width of the SVG element (optional)}

\item{height}{The height of the SVG element (optional)}

\item{units}{The units for width and height (optional)}
}
\value{
An \code{htmltools::HTML()} object
}
\description{
Inserts the content of the SVG file as inline SVG content, i.e. into the
body of the resulting HTML.
}
\details{
In Quarto and R Markdown documents inserting SVG images via standard image
markdown (e.g. \verb{![alt_text](img.svg)}) or HTML \verb{<img>} tags results in an
SVG that is rendered as an image and cannot be interacted with by the
user (e.g. selecting text), nor can elements be modified by CSS.
\code{inline_svg_file()} takes the content of an SVG file and inserts it as
inline SVG content within the body of the rendered HTML document.

Optionally an element \code{id} can be specified (if not provided one will be
created), if provided will overwrite any existing 'id' attribute encoded in
the SVG object.

A \code{source_note} can be provided that will be included after the image, it
has the class "svg-source-note" for styling with CSS and an id starting
with "svg-source-note-" and then using the supplied or generated id.

\code{svglite::svglite()} encodes the width and height (in points) of the object
into the \verb{<svg>} tag (in addition to the viewBox declaration), this can
cause issues for responsive scaling of plots. The \code{dimensions} argument can
\code{"preserve"} any existing dimensions (the default), \code{"strip"} the dimensions
removing them from the tag, or \code{"replace"} them with new dimensions supplied
using the \code{width}, \code{height} and \code{units} arguments.
}
