% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_svg.R
\name{plot}
\alias{plot}
\alias{svg_plot}
\alias{inline_svg_plot}
\title{Render plots as SVG and insert as inline content}
\usage{
svg_plot(
  plot,
  width,
  height,
  units = c("px", "pt", "cm", "mm", "in"),
  path = NULL,
  overwrite = TRUE,
  standalone = TRUE,
  ...
)

inline_svg_plot(
  plot,
  width,
  height,
  units = c("px", "pt", "cm", "mm", "in"),
  path = NULL,
  overwrite = NULL,
  alt_title = NULL,
  alt_description = NULL,
  id = NULL,
  explicit_size = FALSE,
  standalone = FALSE,
  ...
)
}
\arguments{
\item{plot}{A graphics object to plot.}

\item{width}{Width of the output plot.}

\item{height}{Height of the output plot.}

\item{units}{Units used for width and height.}

\item{path}{A file path to save the SVG plot to (optional).}

\item{overwrite}{Should files be overwritten (default is \code{TRUE})}

\item{standalone}{Whether \code{svglite::svglite()} should omit the XML DTD
header and namespace info (default is \code{FALSE}).}

\item{...}{Arguments passed on to \code{svglite::svglite()}.}

\item{alt_title}{Short alt text to include in the inline SVG (optional).}

\item{alt_description}{Long-form alt text to include in the inline SVG
(optional).}

\item{id}{An id for referencing the SVG within the HTML/CSS (optional)}

\item{explicit_size}{Whether the inline \verb{<svg>} tag should explicitly
include the dimensions of the plot (default is \code{FALSE}).}
}
\description{
A wrapper to \code{svglite::svglite()} for saving plots as SVG files and
inserting as inline_content.
}
\details{
\code{svg_plot()} is a wrapper around the \code{svglite::svglite()} graphics
device that simplifies its use for saving plots. Compared to
\code{svglite::svglite()} users do not need to manually active and deactivate
the graphics device, users can also specify width and height in dimensions
other than inches.

\code{inline_svg_plot()} combines \code{svg_plot()} with \code{inline_svg_file()} to insert
the plot as inline SVG content within a Quarto or R Markdown document.

Unless otherwise specified the chart is rendered to SVG via a temporary
file, this path is returned invisibly from \code{svg_plot()}. If re-using plots
elsewhere it is advised to use an explicit path name and use
\code{inline_svg_file()} to re-insert the SVG content rather than rely on the
temporary path provided by \code{svg_plot()}.

An \code{id} for the inline SVG can be provided for so that it is directly
accessible via HTML or CSS (e.g. for linking or styling).
}
\section{Alt text}{
The \code{alt_title} and \code{alt_description} allow for the insertion of alt text
into the SVG content, \code{alt_title} is intended for short descriptions and
\code{alt_description} for longer-form descriptions. If the supplied plot is a
a \code{ggplot2::ggplot()} object and no alt text is supplied then the plot
object will be inspected for alt text, if the alt text is 150 characters or
less it will be used for \code{alt_title} but if longer it will be used for
\code{alt_description}.
}

